chat-server/pyproject.toml
[tool.poetry]
name = "chat-server"
version = "0.1.0"
description = ""
authors = ["yeonwlee <tlstprl1806@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
websockets = "^13.1"
python-dotenv = "^1.0.1"
openai = "^1.52.2"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
